+++ tests/PactNet.Tests/Matchers/Type/TypeMatcherTests.cs
@@ -1,6 +1,5 @@
 using FluentAssertions;
 using Newtonsoft.Json;
-using PactNet.Matchers.Type;
 using Xunit;
 
 namespace PactNet.Tests.Matchers.Type
